requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-core==1.38.2
azure-storage-blob==12.28.0
bcrypt==3.2.2
black==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.6
greenlet==3.3.1
h11==0.16.0
idna==3.11
isodate==0.7.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pypdfium2==5.5.0
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
requests==2.32.5
rsa==4.9.1
ruff==0.15.1
six==1.17.0
SQLAlchemy==2.0.46
sqlmodel==0.0.32
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
